Added the extra cmake options

+++ llvm/scripts/compile.sh
@@ -9,9 +9,11 @@ gitDir="`git rev-parse --show-toplevel`" ;
 LLVM_DIR=`realpath $1` ;
 
 # Create the vanilla LLVM
-git clone https://github.com/scampanoni/LLVM_installer.git $LLVM_VERSION ;
-cd ${LLVM_DIR} ;
-make src 
+if ! test -d $LLVM_VERSION ; then
+  git clone https://github.com/scampanoni/LLVM_installer.git $LLVM_VERSION ;
+  cd ${LLVM_DIR} ;
+  make src 
+fi
 
 # Apply the patch
 cd ${LLVM_DIR}/src ; 
@@ -19,8 +21,10 @@ tar xf $patchFile ;
 
 # Link the clang plugin
 cd ${LLVM_DIR}/src/tools/clang/tools ; 
-ln -s ${gitDir}/clang-tools hobbit-tricks ;
+if ! test -e hobbit-tricks ; then
+  ln -s ${gitDir}/clang-tools hobbit-tricks ;
+fi
 
 # Compile
 cd ${LLVM_DIR}; 
-make EXTRAS="noextra" TESTS="notest" ;
+make EXTRAS="noextra" TESTS="notest" EXTRA_CMAKE_OPTIONS="-DCMAKE_CXX_COMPILER=clang++ -DLLVM_BUILD_LLVM_DYLIB=ON -DLLVM_LINK_LLVM_DYLIB=ON";
